Added support to download and upload the filesystem image.
Downloading will provide you with a raw disk image; just like a dd if=/dev/sda of=/diskimg. Only this download will be a lot smaller ;)

There is also a button to clear the local storage. Press it, and immediatly refresh to start over on a completely empty localstorage (for this origin at least).

The last added option is to upload a stored disk image, so you can resume with your data. Remember that this is a fixed size disk so larger or smaller images can't be used, they require a driver update.